requirements.txt
google-cloud-storage
google-cloud-aiplatform
pandas
matplotlib
streamlit
gradio
dvc
mlflow
chromadb